combined token/auth intercepter cleaned and corrected
