+++ outreachHub_angular/outreachHub_user/src/app/core/core.module.ts
@@ -1,6 +1,6 @@
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { AuthInterceptor } from './interceptors/auth.interceptor';
+import { AuthInterceptor } from './auth.interceptor';
 
 @NgModule({
   imports: [HttpClientModule],
